+++ src/lib/components/NavigationBar/NavigationBar.svelte
@@ -9,10 +9,14 @@
     position: fixed;
     top: 0;
     left: 0;
+
     width: 100%;
     height: $navigation-bar-height;
+
     background-color: #1f1f1f;
     color: white;
+
     display: flex;
+    z-index: 10;
   }
 </style>