+++ src/lib/workers/basisTypes.d.ts
@@ -1,5 +1,8 @@
-type MessageFromWorker = EncodingMessageFromWorker | ErrorMessageFromWorker;
-type MessageToWorker = EncodingMessageToWorker;
+type MessageFromWorker =
+  | EncodingMessageFromWorker
+  | TranscodingMessageFromWorker
+  | ErrorMessageFromWorker;
+type MessageToWorker = EncodingMessageToWorker | TranscodingMessageToWorker;
 
 type EncodingMessageToWorker = {
   mode: "encode";
@@ -10,9 +13,22 @@ type EncodingMessageToWorker = {
   }[];
 };
 
+type TranscodingMessageToWorker = {
+  mode: "transcode";
+  basisEncoderJS: string;
+  data: { name: string; file: Blob }[];
+  canvas: OffscreenCanvas;
+};
+
 type EncodingMessageFromWorker = {
   name: string;
   data: Uint8Array;
+  status: "success";
+};
+
+type TranscodingMessageFromWorker = {
+  name: string;
+  blobURL: string;
   status: "success";
 };
 
@@ -24,6 +40,8 @@ type ErrorMessageFromWorker = {
 export {
   EncodingMessageToWorker,
   EncodingMessageFromWorker,
+  TranscodingMessageToWorker,
+  TranscodingMessageFromWorker,
   MessageFromWorker,
   MessageToWorker,
   ErrorMessageFromWorker,